Add a distance function

+++ src/irdm/descriptors/Descriptor.java
@@ -9,7 +9,7 @@ public interface Descriptor {
 
     }
 
-    public double distance(Descriptor o2);
+    public int distance(Descriptor o2);
 
     public static double distance(Descriptor o1, Descriptor o2) {
         return o1.distance(o2);
